Show platform not supported warning only on plugin creation (#6801)
