+++ plugins/inputs/ethtool/ethtool_notlinux.go
@@ -3,11 +3,14 @@
 package ethtool
 
 import (
-	"log"
-
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
+
+func (e *Ethtool) Init() error {
+	e.Log.Warn("Current platform is not supported")
+	return nil
+}
 
 func (e *Ethtool) Gather(acc telegraf.Accumulator) error {
 	return nil
@@ -15,7 +18,6 @@ func (e *Ethtool) Gather(acc telegraf.Accumulator) error {
 
 func init() {
 	inputs.Add(pluginName, func() telegraf.Input {
-		log.Print("W! [inputs.ethtool] Current platform is not supported")
 		return &Ethtool{}
 	})
 }
