+++ plugins/inputs/synproxy/synproxy_notlinux.go
@@ -3,29 +3,21 @@
 package synproxy
 
 import (
-	"log"
-
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
 
-type Synproxy struct{}
+func (k *Synproxy) Init() error {
+	k.Log.Warn("Current platform is not supported")
+	return nil
+}
 
 func (k *Synproxy) Gather(acc telegraf.Accumulator) error {
 	return nil
 }
 
-func (k *Synproxy) Description() string {
-	return ""
-}
-
-func (k *Synproxy) SampleConfig() string {
-	return ""
-}
-
 func init() {
 	inputs.Add("synproxy", func() telegraf.Input {
-		log.Print("W! [inputs.synproxy] Current platform is not supported")
 		return &Synproxy{}
 	})
 }
