+++ plugins/inputs/wireless/wireless_notlinux.go
@@ -3,11 +3,14 @@
 package wireless
 
 import (
-	"log"
-
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
+
+func (w *Wireless) Init() error {
+	w.Log.Warn("Current platform is not supported")
+	return nil
+}
 
 func (w *Wireless) Gather(acc telegraf.Accumulator) error {
 	return nil
@@ -15,7 +18,6 @@ func (w *Wireless) Gather(acc telegraf.Accumulator) error {
 
 func init() {
 	inputs.Add("wireless", func() telegraf.Input {
-		log.Print("W! [inputs.wireless] Current platform is not supported")
 		return &Wireless{}
 	})
 }
